refactor(pool): move null pool base to its own module
And some more light refactoring.
